Resolved change requests
