+++ server/utils/embedUtils.ts
@@ -1,3 +1,4 @@
+import { JSDOM } from "jsdom";
 import rehypePresetMinify from "rehype-preset-minify";
 import rehypeStringify from "rehype-stringify";
 import remarkParse from "remark-parse";
@@ -5,6 +6,7 @@ import remarkRehype from "remark-rehype";
 import { unified } from "unified";
 
 import { Page } from "~/types/server";
+import { EmbedConfigurations } from "~/types/server/nextApi.type";
 
 import { PAGES_COLLECTION } from "../firebase/firestoreCollections";
 import CustomApiError from "./errors/customApiError";
@@ -28,4 +30,22 @@ export async function md2html(md: string) {
         .use(rehypePresetMinify)
         .use(rehypeStringify);
     return String(await processor.process(md));
+}
+
+export function generateCommentHTML(html: string, config: EmbedConfigurations, isDark?: boolean) {
+    if (typeof window !== "undefined")
+        throw new CustomApiError("This function should be ran on the server", 403);
+    const dom = new JSDOM(html);
+    const document = dom.window.document;
+    const script = document.createElement("script");
+    script.textContent = `
+        import ezkomment from "/v1/js/${
+            process.env.NODE_ENV === "production" ? "ezkomment.min.js" : "ezkomment.js"
+        }";
+        ezkomment(${JSON.stringify(config)});
+    `;
+    script.setAttribute("type", "module");
+    document.head.appendChild(script);
+    if (isDark) document.documentElement.classList.add("dark");
+    return dom.serialize();
 }
